consumers/analytics-service/requirements.txt
kafka-python==2.0.2
couchbase==4.3.0
